Module 1-Env.Block_header

type shell_header = {

level : Int32.t;
The number of preceding block in this chain, i.e. the genesis has level 0.

proto_level : int;
The number of preceding protocol change in the chain (modulo 256), i.e. the genesis has proto_level 0.

predecessor : Block_hash.t;

timestamp : Time.t;

validation_passes : int;

operations_hash : Operation_list_list_hash.t;

fitness : Bytes.t list;

context : Context_hash.t;

}

val shell_header_encoding : shell_header Data_encoding.t

type t = {

shell : shell_header;

protocol_data : bytes;

}

include S.HASHABLE with type t := t and type hash := Block_hash.t

include S.T

type t

include Compare.S with type t := t

type t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val pp : Format.formatter -> t -> unit

val encoding : t Data_encoding.t

val to_bytes : t -> bytes

val of_bytes : bytes -> t option

type hash

val hash : t -> hash

val hash_raw : bytes -> hash